feat: changes.

+++ pkg/storage/evictor.go
@@ -110,7 +110,10 @@ func (e *Evict) evictUntilWithinLimit() (items int, mem int64) {
 		for e.shouldEvictRightNow() {
 			el, ok := shard.LruList().Next(offset)
 			if !ok {
-				break
+				break // end of the LRU list, move to next
+			}
+			if !el.Value().Acquire() {
+				continue // already marked as doomed but not removed yet, skip it
 			}
 
 			freedMem, isHit := e.db.Remove(el.Value())
@@ -120,6 +123,7 @@ func (e *Evict) evictUntilWithinLimit() (items int, mem int64) {
 				mem += freedMem
 			}
 
+			el.Value().Release()
 			offset++
 		}
 	}
